+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 - Add an example script demonstrating how tc can use external contraction path finder wirtten in Julia
 
+- Add `cals_from_api` method for `ReadoutMit` class which can acquire the readout error information from the api
+
 ### Fixed
 
 - Circuit nosify in noise model now support all circuit attributs apart from qubit number